# Lazy loading models for faster convert and merge