--- a/llms/mlx_lm/utils.py
+++ b/llms/mlx_lm/utils.py
@@ -1,4 +1,5 @@
 import copy
+import gc
 import glob
 import importlib
 import json
@@ -467,6 +468,7 @@ def save_weights(
     # necessary ones
     if donate_weights:
         weights.clear()
+        gc.collect()
 
     for i in range(len(shards)):
         shard = shards[i]
@@ -479,6 +481,7 @@ def save_weights(
         for weight_name in shard.keys():
             index_data["weight_map"][weight_name] = shard_name
         del shard
+        gc.collect()
 
     index_data["weight_map"] = {
         k: index_data["weight_map"][k] for k in sorted(index_data["weight_map"])