Currently `load_model` in utils immediately evaluates the parameters which may be far from ideal for `merge` or `convert`. This adds the optional `lazy` parameter to `fetch_from_hub` and `load` as well as `donate_weights` to `save_weights` which allows us to only partially load a model when doing the above operations.

Sorry that you 'll have to scroll right to see the times. TL;DR: When memory is scarce things get much faster. I will report back with results from quantizing Mixtral on my puny M2 air.

Before

```
$ python -m mlx_lm.merge --config llms/mlx_lm/examples/merge_config.yaml    4.68s user 28.93s system 32% cpu 1:42.12 total
$ python -m mlx_lm.convert --hf-path /tmp/merged --mlx-path /tmp/merged-q4-test  2.64s user 9.67s system 43% cpu 27.997 total
$ python -m mlx_lm.convert --hf-path /tmp/merged --mlx-path /tmp/merged-q4-test  4.74s user 30.25s system 68% cpu 51.430 total
```

After

```
$ python -m mlx_lm.merge --config llms/mlx_lm/examples/merge_config.yaml    3.48s user 18.11s system 86% cpu 25.037 total
$ python -m mlx_lm.convert --hf-path /tmp/merged --mlx-path /tmp/merged-q4-test  2.59s user 7.26s system 41% cpu 23.462 total
$ python -m mlx_lm.convert --hf-path /tmp/merged --mlx-path /tmp/merged-q4-test  4.58s user 28.26s system 76% cpu 42.790 total
```

Side note: I left this on top of the model merging PR so the diff may be a bit annoying at the moment. I will rebase on top of main when that is merged.