@awni 2024-02-20
Also, what's the model you reported times for? 7B mistral? Did the Mixtral model actually run?