@angeloskath 2024-02-20
So just ran Mixtral quantization on my laptop with 24GB of RAM (for the 2nd time cause the 1st it ran in the background and I didn't notice it).

It takes 3 minutes which is quite ok provided it has to read ~100GB and write ~25GB which at 1GB/s is about 2 minutes without doing any processing whatsoever (also I don't think my disk can sustain 1GB/s for 120GB).

Otoh it did swap more than a bit. Something like 17GB of swap I am not sure why that is. A huge improvement from being killed 1 minute in without lazy loading but still. Will do a bit of investigating but this should be good to merge regardless after rebase .